styling improved, pie chart complete

+++ db/migrate/20230506183921_create_allowances.rb
@@ -1,7 +1,7 @@
 class CreateAllowances < ActiveRecord::Migration[7.0]
   def change
     create_table :allowances do |t|
-      t.integer :balance
+      t.float :balance
       t.integer :kid_id
       t.integer :parent_id
